openclaw-bridge/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv==1.1.1
azure-identity>=1.19.0
